+++ raaz-core/Raaz/Primitive/Blake2s/Implementation/CPortable.hs
@@ -2,8 +2,8 @@
 {-# LANGUAGE DataKinds                  #-}
 {-# LANGUAGE KindSignatures             #-}
 
--- | The portable C-implementation of Blake2b.
-module Raaz.Primitive.Blake2b.CPortable where
+-- | The portable C-implementation of Blake2s.
+module Raaz.Primitive.Blake2s.Implementation.CPortable where
 
 import Foreign.Ptr                ( Ptr          )
 import Control.Monad.IO.Class     ( liftIO       )
@@ -19,68 +19,64 @@ import Raaz.Primitive.Blake2.Internal
 
 
 name :: String
-name = "blake2b-cportable"
+name = "blake2s-cportable"
 
 description :: String
-description = "BLAKE2b Implementation using portable C and Haskell FFI"
+description = "BLAKE2s Implementation using portable C and Haskell FFI"
 
-type Prim                    = BLAKE2b
-type Internals               = Blake2bMem
+type Prim                    = BLAKE2s
+type Internals               = Blake2sMem
 type BufferAlignment         = 32
 
 
-additionalBlocks :: BLOCKS BLAKE2b
+additionalBlocks :: BLOCKS Prim
 additionalBlocks = blocksOf 1 Proxy
 
+------------------------- FFI For Blake2s -------------------------------------
 
------------------------- The foreign function calls  ---------------------
 
 foreign import ccall unsafe
-  "raaz/hash/blake2/common.h raazHashBlake2bPortableBlockCompress"
-  c_blake2b_compress  :: AlignedPointer BufferAlignment
+  "raaz/hash/blake2/common.h raazHashBlake2sPortableBlockCompress"
+  c_blake2s_compress  :: Pointer
                       -> Int
-                      -> Ptr (BYTES Word64)
-                      -> Ptr (BYTES Word64)
-                      -> Ptr BLAKE2b
+                      -> BYTES Word64
+                      -> Ptr Prim
                       -> IO ()
 
 foreign import ccall unsafe
-  "raaz/hash/blake2/common.h raazHashBlake2bPortableLastBlock"
-  c_blake2b_last   :: Pointer
+  "raaz/hash/blake2/common.h raazHashBlake2sPortableLastBlock"
+  c_blake2s_last   :: Pointer
                    -> BYTES Int
                    -> BYTES Word64
-                   -> BYTES Word64
-                   -> Word64
-                   -> Word64
-                   -> Ptr BLAKE2b
+                   -> Word32
+                   -> Word32
+                   -> Ptr Prim
                    -> IO ()
-
 --
 processBlocks :: AlignedPointer BufferAlignment
-              -> BLOCKS BLAKE2b
-              -> MT Blake2bMem ()
+              -> BLOCKS Prim
+              -> MT Internals ()
 
-processBlocks buf blks = do uPtr   <- uLengthCellPointer
-                            lPtr   <- lLengthCellPointer
-                            hshPtr <- hashCell128Pointer
-                            liftIO $ c_blake2b_compress buf (fromEnum blks) uPtr lPtr hshPtr
+processBlocks buf blks =
+  do l      <- getLength
+     hashCellPointer >>= liftIO . c_blake2s_compress (forgetAlignment buf) (fromEnum blks) l
+     updateLength blks
 
 -- | Process the last bytes.
 processLast :: AlignedPointer BufferAlignment
             -> BYTES Int
-            -> MT Blake2bMem ()
+            -> MT Internals ()
 processLast buf nbytes  = do
   unsafeWrite padding $ forgetAlignment buf  -- pad the message
   processBlocks buf nBlocks                  -- process all but the last block
   --
   -- Handle the last block
   --
-  u      <- getULength
-  l      <- getLLength
-  hshPtr <- hashCell128Pointer
-  liftIO $ c_blake2b_last lastBlockPtr remBytes u l f0 f1 hshPtr
+  l      <- getLength
+  hshPtr <- hashCellPointer
+  liftIO $ c_blake2s_last lastBlockPtr remBytes l f0 f1 hshPtr
 
-  where padding      = blake2Pad (Proxy :: Proxy BLAKE2b) nbytes
+  where padding      = blake2Pad (Proxy :: Proxy Prim) nbytes
         nBlocks      = atMost (bytesToWrite padding) <> toEnum (-1) -- all but the last block
         remBytes     = nbytes - inBytes nBlocks                     -- Actual bytes in the last block.
         lastBlockPtr = forgetAlignment buf `movePtr` nBlocks
