+++ raaz-core/Raaz/Primitive/Sha512/Implementation/CPortable.hs
@@ -2,8 +2,8 @@
 {-# LANGUAGE KindSignatures             #-}
 {-# LANGUAGE DataKinds                  #-}
 
--- | The portable C-implementation of SHA256.
-module Raaz.Primitive.Sha256.CPortable
+-- | The portable C-implementation of SHA512.
+module Raaz.Primitive.Sha512.Implementation.CPortable
        ( name, description
        , Prim, Internals, BufferAlignment
        , additionalBlocks
@@ -20,63 +20,67 @@ import Data.Proxy
 
 import Raaz.Core
 import Raaz.Primitive.HashMemory
-import Raaz.Primitive.Sha256.Internal
+import Raaz.Primitive.Sha512.Internal
 
 
 name :: String
-name = "sha256-cportable"
+name = "sha512-cportable"
 
 description :: String
-description = "SHA256 Implementation using portable C and Haskell FFI"
+description = "SHA512 Implementation using portable C and Haskell FFI"
 
-type Prim                    = SHA256
-type Internals               = Sha256Mem
+type Prim                    = SHA512
+type Internals               = Sha512Mem
 type BufferAlignment         = 32
 
 
-additionalBlocks :: BLOCKS SHA256
+additionalBlocks :: BLOCKS SHA512
 additionalBlocks = blocksOf 1 Proxy
 
 ------------------------ The foreign function calls  ---------------------
 
 foreign import ccall unsafe
-  "raaz/hash/sha256/portable.h raazHashSha256PortableCompress"
-  c_sha256_compress  :: Pointer -> Int -> Ptr SHA256 -> IO ()
+  "raaz/hash/sha512/portable.h raazHashSha512PortableCompress"
+  c_sha512_compress  :: Pointer -> Int -> Ptr SHA512 -> IO ()
 
 
 compressBlocks :: AlignedPointer BufferAlignment
-               -> BLOCKS SHA256
+               -> BLOCKS SHA512
                -> MT Internals ()
-compressBlocks buf blks =  hashCellPointer >>= liftIO . c_sha256_compress (forgetAlignment buf) (fromEnum blks)
+compressBlocks buf blks =  hashCell128Pointer
+                           >>= liftIO . c_sha512_compress (forgetAlignment buf) (fromEnum blks)
 
 
 processBlocks :: AlignedPointer BufferAlignment
-              -> BLOCKS SHA256
+              -> BLOCKS SHA512
               -> MT Internals ()
-processBlocks buf blks = compressBlocks buf blks >> updateLength blks
+processBlocks buf blks = compressBlocks buf blks >> updateLength128 blks
 
 
 
 -- | Padding is message followed by a single bit 1 and a glue of zeros
 -- followed by the length so that the message is aligned to the block boundary.
 padding :: BYTES Int    -- Data in buffer.
-        -> BYTES Word64 -- Message length
+        -> BYTES Word64 -- Message length higher
+        -> BYTES Word64 -- Message length lower
         -> WriteM (MT Internals)
-padding bufSize msgLen  = glueWrites 0 boundary hdr lengthWrite
+padding bufSize uLen lLen  = glueWrites 0 boundary hdr lengthWrite
   where skipMessage = skipWrite bufSize
         oneBit      = writeStorable (0x80 :: Word8)
         hdr         = skipMessage <> oneBit
-        boundary    = blocksOf 1 (Proxy :: Proxy SHA256)
-        lengthWrite = write $ bigEndian (shiftL w 3)
-        BYTES w     = msgLen
+        boundary    = blocksOf 1 (Proxy :: Proxy SHA512)
+        lengthWrite = write (bigEndian up) <> write (bigEndian lp)
+        BYTES up    = shiftL uLen 3 .|. shiftR lLen 61
+        BYTES lp    = shiftL lLen 3
 
 -- | Process the last bytes.
 processLast :: AlignedPointer BufferAlignment
             -> BYTES Int
             -> MT Internals ()
 processLast buf nbytes  = do
-  updateLength nbytes
-  totalBytes  <- getLength
-  let pad      = padding nbytes totalBytes
+  updateLength128 nbytes
+  uLen  <- getULength
+  lLen  <- getLLength
+  let pad      = padding nbytes uLen lLen
       blocks   = atMost $ bytesToWrite pad
       in unsafeWrite pad (forgetAlignment buf) >> compressBlocks buf blocks
