+++ raaz-core/implementation/cbits/raaz/cipher/aes/common.c
