+++ raaz-core/implementation/cbits/raaz/cipher/aes/cportable.c
