+++ raaz-core/implementation/cbits/raaz/cipher/chacha20/cportable.c
