+++ raaz-core/implementation/cbits/raaz/cipher/cportable/salsa20.c
