+++ raaz-core/implementation/cbits/raaz/hash/blake2/blake2b/portable.c
