+++ raaz-core/implementation/cbits/raaz/hash/blake2/common.h
