+++ raaz-core/implementation/cbits/raaz/hash/sha1/portable.c
