+++ raaz-core/implementation/cbits/raaz/hash/sha256/portable.c
