+++ raaz-core/implementation/cbits/raaz/hash/sha512/portable.c
