+++ raaz-core/raaz-core.cabal
@@ -139,3 +139,21 @@ library
   include-dirs: cbits
   includes: raaz/core/endian.h
   install-includes: cbits/raaz/core/endian.h
+  hs-source-dirs: implementation
+  exposed-modules: Raaz.Primitive.Blake2b.Implementation.CPortable
+                 , Raaz.Primitive.Blake2s.Implementation.CPortable
+                 , Raaz.Primitive.Sha256.Implementation.CPortable
+                 , Raaz.Primitive.Sha512.Implementation.CPortable
+                 , Raaz.Primitive.ChaCha20.Implementation.CPortable
+
+  c-sources: implementation/cbits/raaz/hash/blake2/blake2b/portable.c
+           , implementation/cbits/raaz/hash/blake2/blake2s/portable.c
+           , implementation/cbits/raaz/hash/sha1/portable.c
+           , implementation/cbits/raaz/hash/sha256/portable.c
+           , implementation/cbits/raaz/hash/sha512/portable.c
+           , implementation/cbits/raaz/cipher/chacha20/cportable.c
+  include-dirs: implementation/cbits/
+  includes: implementation/cbits/raaz/hash/blake2/common.h
+          , implementation/cbits/raaz/hash/blake2/blake2b/constants.h
+          , implementation/cbits/raaz/hash/blake2/blake2s/constants.h
+
